Exception handling

+++ src/main/java/com/guilherme/crud_client/services/ClientService.java
@@ -3,10 +3,17 @@ package com.guilherme.crud_client.services;
 import com.guilherme.crud_client.dto.ClientDTO;
 import com.guilherme.crud_client.entities.Client;
 import com.guilherme.crud_client.repositories.ClientRepository;
+import com.guilherme.crud_client.services.exceptions.DatabaseException;
+import com.guilherme.crud_client.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -18,8 +25,9 @@ public class ClientService {
     private ClientRepository repository;
 
     @Transactional(readOnly = true)
-    public ClientDTO findById(Long id) {
-            Client client = repository.findById(id).get();
+        public ClientDTO findById(Long id) {
+            Client client = repository.findById(id).orElseThrow(
+                    ()-> new ResourceNotFoundException("Recurso Não Encontrado"));
             return new ClientDTO(client);
     }
 
@@ -40,16 +48,29 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto) {
+        try {
+            Client entity = repository.getReferenceById(id);
+            copyDtoEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ClientDTO(entity);
+        }
+        catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
 
-        Client entity = repository.getReferenceById(id);
-        copyDtoEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ClientDTO(entity);
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try{
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoEntity(ClientDTO dto, Client entity) {
@@ -59,5 +80,5 @@ public class ClientService {
         entity.setBirthDate(dto.getBirthDate());
         entity.setChildren(dto.getChildren());
     }
-    
+
 }
